Use Double precision instead of float

+++ RTree/Point.cs
@@ -39,7 +39,7 @@ namespace RTree
         /// <summary>
         /// The (x, y) coordinates of the point.
         /// </summary>
-        internal float[] coordinates;
+        internal double[] coordinates;
 
 
         /// <summary>
@@ -48,9 +48,9 @@ namespace RTree
         /// <param name="x">The x coordinate of the point</param>
         /// <param name="y">The y coordinate of the point</param>
         /// <param name="z">The z coordinate of the point</param>
-        public Point(float x, float y,float z)
+        public Point(double x, double y,double z)
         {
-            coordinates = new float[DIMENSIONS];
+            coordinates = new double[DIMENSIONS];
             coordinates[0] = x;
             coordinates[1] = y;
             coordinates[2] = z;
